Properly checkpoint stateful nodes
  Properly checkpoint stateful nodes (Dropout, RandomSample) in both v1/v2
  formats:
     To ensure reproducibility, nodes that inherit from RngUser need to
  properly save and restore their internal stated (rng seed and offset).
     V2 additionally needs to copy state from the checkpoint into the function
  graph and the underlying network if it already exists.

+++ Source/Common/Include/fileutil.h
@@ -484,7 +484,7 @@ const wchar_t* GetFormatString(float);
 template <>
 const wchar_t* GetFormatString(double);
 template <>
-const wchar_t* GetFormatString(size_t);
+const wchar_t* GetFormatString(unsigned long long);
 template <>
 const wchar_t* GetFormatString(long long);
 template <>
@@ -522,7 +522,7 @@ const wchar_t* GetScanFormatString(float);
 template <>
 const wchar_t* GetScanFormatString(double);
 template <>
-const wchar_t* GetScanFormatString(size_t);
+const wchar_t* GetScanFormatString(unsigned long long);
 template <>
 const wchar_t* GetScanFormatString(long long);
